add api namespaces and split them into modules
