+++ api/.flaskenv
@@ -1,2 +1,2 @@
-FLASK_APP=main.py
+FLASK_APP=__init__.py
 FLASK_ENV=development
